requirements.txt
matplotlib==3.5.3
openai==0.23.0
Pillow==9.2.0
python-dotenv==0.21.0
requests==2.27.1
